Replaced a few portions of code with calls to memcpy(), strncpy()

+++ compress.c
@@ -257,7 +257,7 @@ void sendMTFValues ( EState* s )
 
    alphaSize = s->nInUse + 2;
    for (t = 0; t < BZ_N_GROUPS; t++)
-      memset (s->len[t], BZ_GREATER_ICOST, alphaSize);
+      memset (s->len[t], BZ_GREATER_ICOST, (size_t)alphaSize);
 
    /*--- Decide how many coding tables to use ---*/
    AssertH ( s->nMTF > 0, 3001 );
@@ -315,7 +315,7 @@ void sendMTFValues ( EState* s )
       memset (fave, 0, (size_t)nGroups * sizeof(int32_t));
 
       for (t = 0; t < nGroups; t++)
-         memset (s->rfreq[t], 0, alphaSize * sizeof(int32_t));
+         memset (s->rfreq[t], 0, (size_t)alphaSize * sizeof(int32_t));
 
       /*---
         Set up an auxiliary length table which is used to fast-track
@@ -343,7 +343,7 @@ void sendMTFValues ( EState* s )
             Calculate the cost of this group as coded
             by each of the coding tables.
          --*/
-         memset (cost, 0, nGroups * sizeof(uint16_t));
+         memset (cost, 0, (size_t)nGroups * sizeof(uint16_t));
 
          if (nGroups == 6 && 50 == ge-gs+1) {
             /*--- fast track the common case ---*/
